Optional Lights
Using lights will reduce performance. Additionally, there is a limit to how many can be rendered at once.
It is also possible to fake lights by using a Sprite2D and setting it's mix mode to 'Add', but this method seems to look more musty and will illuminate the environment background as well.